+++ src/client/src/containers/DependencyInfo/index.tsx
