+++ src/client/src/containers/DependencyInfo/styles.module.css
